miglioramenti post
Activity Test.class (Creato solo per testare i post)

+++ 3_CodiceSorgente/IngSW_NaTour21/app/src/main/java/com/example/natour21/ParentItem.java
@@ -3,9 +3,9 @@ package com.example.natour21;
 public class ParentItem {
 
     String nomePercorso, difficolta, tempo, area;
-    int rate;
+    float rate;
 
-    public ParentItem(String nomePercorso, String difficolta, String tempo, String area, int rate) {
+    public ParentItem(String nomePercorso, String difficolta, String tempo, String area, float rate) {
         this.nomePercorso = nomePercorso;
         this.difficolta = difficolta;
         this.tempo = tempo;
